Remove renderer method. Fix typo.
